feat(#180): 목표 Summary UI 에 Error / Loading UI 추가
